ecosystem_integration_templates/teams_datarobot/requirements.txt
botbuilder-integration-aiohttp>=4.14.0
datarobotx
datarobot
textstat
tiktoken
nltk
py-readability-metrics
datarobot-mlops
datarobot-mlops-connected-client